Show Kubernetes events in UI
