+++ lib/monitoring/addon/components/cluster-dashboard-tabs/template.hbs
@@ -39,4 +39,11 @@
   }}
     {{rancher-metrics largeScale=largeScale noDataLabel="clusterDashboard.noRancherComponents"}}
   {{/metrics-summary}}
+
+  {{resource-event-list
+    resourceType=(t "generic.cluster")
+    clusterEvents=true
+    expandAll=al.expandAll
+    expandFn=expandFn
+  }}
 {{/accordion-list}}
