+++ lib/monitoring/addon/components/cluster-dashboard/template.hbs
@@ -99,6 +99,15 @@
 
   {{#if (and showClusterTabs scope.currentCluster.isMonitoringReady)}}
     {{cluster-dashboard-tabs dashboards=dashboards}}
+  {{else}}
+    {{#accordion-list showExpandAll=false as | al expandFn |}}
+      {{resource-event-list
+        resourceType=(t "generic.cluster")
+        clusterEvents=true
+        expandAll=al.expandAll
+        expandFn=expandFn
+      }}
+    {{/accordion-list}}
   {{/if}}
 {{else}}
   <div class="text-center text-muted">
