+++ lib/shared/app/components/resource-event-list/component.js
@@ -1,0 +1,1 @@
+export { default } from 'shared/components/resource-event-list/component';
