+++ translations/en-us.yaml
@@ -3548,9 +3548,11 @@ resourceEventList:
   detail: Events of current {resourceType}
   table:
     type: Type
-    lastUpdate: Last Update
+    lastUpdate: Last Seen
     reason: Reason
     message: Message
+    namespace: Namespace
+    object: Object
   noData: No Events
   noMatch: No events match the current search
